pyproject.toml
[tool.poetry]
name = "yet-another-aliyun-exporter"
version = "0.2.3"
description = "Yet another aliyun exporter"
authors = ["zaldnoay <zunway@outlook.com>"]
license = "LGPL-2.1-or-later"
readme = "README.md"
packages = [
    { include = "aliyun_exporter" }
]

[tool.poetry.dependencies]
python = ">=3.7.1"
marshmallow = "^3"
"ruamel.yaml" = "^0.17"
alibabacloud-cms20190101 = "^2"
prometheus-client = "^0.11"
alibabacloud-tag20180828 = "^1"
Twisted = "^21"
click = "^8"

[tool.poetry.dev-dependencies]
ipykernel = "^6"
pandas = "^1"
nbconvert = "^6"

[tool.poetry.scripts]
start = "aliyun_exporter:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
